implement an energy cascading result finder
The scripts calculating the energy cascading matrices (for collision with each
neutral species) will now look for already existing matrices in the E_cascadings
folder. If a matrix with matching energy grid is found, the matrix is loaded.
Otherwise, a new matrice matching the energy grid is calculated and saved in the
E_cascadings folder for future uses.

These energy cascading matrix calculations are quite time consuming. Thus the
implementation of this functionality.