set up proper warnings in the make_all_blabla.m files
If the variables 'results_dir' and 'RunDirs' were not existing,
the scripts were automatically attributing them some values.
For 'results_dir', this was leading to the creation of obscure folders
names and places for storing the outputs.
For 'RunDirs' this was simply leading to an error.

I think it is better with some proper warnings remembering the user to
enter something for these variables.