correct the E_cascading finder
The finder was used only if Q was empty. The problem with this is
that if the user run the program and then change the energy grid
for another run, Q is not empty and the finder loop is not used,
leading to unnecessary energy cascading calculations.
This problem is solved by adding a check to the find loop. The finder
is now used if Q is empty OR if E4Q and the energy grid are
different.