improve the E_cascading finder script
First improvement is that the old script was
testing if Q was empty or if the energy grid of
the secondary electrons (Es) was matching the
energy grid with which Q had been computed
(E4Q). This with the following test :

	if isempty(Q) || ~all(E4Q(1:min(numel(Es),numel(E4Q)))==Es(1:min(numel(Es),numel(E4Q)))) || size(Es,2) > size(E4Q,2)

A problem was occuring when Es was changed
between two runs. Q was not empty as the
function had been called in the previous run.
So the first test would pass. If Es was made
bigger (higher incoming energy for exemple) but
the lower end was the same as before, the
second test would match the lower end of Es and
E4Q, and the script would skip the
loading/calculation of the correct cascading
matrix Q. When the function was then trying to
calculate the cascading or secondary electron
spectra, the second index in Q was too big
(remember the new Es being bigger than the one
Q was calculated with). To solve that, a third
test is added which tests if Es is bigger than
E4Q :

	|| size(Es,2) > size(E4Q,2)

Second improvement is that for loading a
preexisting cascading matrix Q, the function
was testing if E4Q and Es were equal with the
following command :

	if isequal (E4Q,Es)

We just saw that Es should not be bigger than
E4Q. But the opposite is not a problem, as long
as the lower end of E4Q is matching Es. The
test above was not testing this possibility and
failed to load some preexisting Q matrix when
Es was smaller than the one Q had been
calculated with (E4Q). This led to some
unnecessary calculations of Q that are
time-consuming. Thus, the new test is :

	if all(E4Q(1:min(numel(Es),numel(E4Q)))==Es)