change the order of N2 ionization collisions in .dat files
The double ionization collision was written
before some other simple ionization collisions
in the .dat files containing the cross sections
and energy of secondary electrons for
collisions of e- with N2. For security reasons
we move that to the end of the file, as it is
already the case for the files related to
collisions with O and O2. This might help avoid
some unexpected breaking of loops or such in the
future.