fix the energy problem! closes #2
The problem was in the CN scheme. It was an issue
with how the matrix for the space differential
factor was constructed. The indexes to use for
the diagonal+1 and diagonal-1 have to be
different as they don't start and stop at same
altitudes. The h4diff array also has to be
different for upstreams and downstreams going e-,
as the scheme is starting on opposite ends of the
altitude atmosphere.

Some other change : half steps in space are
removed. I cannot find any reason for us to use
them in the scheme.