Remove "stepE" part and change way to choose Emax
The "stepE" part is not neccesary as it was commented away in
setup4etrp.m, so simulations are always computed on a stepE = 1 grid.

About the Emax, the user now enters a maximum energy in eV instead of
an index iEmax. The program finds iEmax itself from the given Emax.