Update Loop_MI_coupling to work with dt = 1e-3s
Change the t_run method to always have time slices of the same length.

Also comment away the choices of n_t and Dir-name parameters as these
should now be chosen in the Coupling_control.m file.

Also deleted the "if setup complete" test so that the setup script is
always being run. This is to avoid problems if several runs with
different parameters are launched in one same session. It is also not a
problem to run the setup a lot of times (compared to at the beginning of
the development of AURORA) as all time-intensive calculations are now
done in functions that check if the calculations have already been done
and loads them if that is the case.